Added new file for Pet class
